+++ src/Service/Route53/CHANGELOG.md
@@ -9,6 +9,8 @@
 - AWS api-change: Added `us-iso-west-1` region
 - AWS api-change: Added `ap-southeast-3` region
 - Fixed parameter trim in the called API urls
+- AWS api-change: Add new APIs to support Route 53 IP Based Routing
+- AWS enhancement: SDK doc update for Route 53 to update some parameters with new information.
 
 ## 0.1.1